+++ src/formats.rs
@@ -225,6 +225,20 @@ formats! {
     extension = "bpg"
     kind = Image
 
+    format = BitmapFontAscii
+    name = "Bitmap Font ASCII"
+    short_name = "FNT"
+    media_type = "font/x-fnt-ascii"
+    extension = "fnt"
+    kind = Font
+
+    format = BitmapFontBinary
+    name = "Bitmap Font Binary"
+    short_name = "FNT"
+    media_type = "font/x-fnt-binary"
+    extension = "fnt"
+    kind = Font
+
     format = BittorrentFile
     name = "BitTorrent File"
     short_name = "Torrent"